Refactor namespaces for Input classes to improve organization and clarity
