+++ src/Input/InputParam.php
@@ -2,8 +2,9 @@
 
 declare(strict_types=1);
 
-namespace BEAR\Resource;
+namespace BEAR\Resource\Input;
 
+use BEAR\Resource\ParamInterface;
 use Ray\Di\InjectorInterface;
 use ReflectionNamedType;
 use ReflectionParameter;
