+++ src/NamedParamMetas.php
@@ -6,6 +6,7 @@ namespace BEAR\Resource;
 
 use BEAR\Resource\Annotation\RequestParamInterface;
 use BEAR\Resource\Annotation\ResourceParam;
+use BEAR\Resource\Input\InputParam;
 use BEAR\Resource\Options\OptionalParam;
 use Ray\Aop\ReflectionMethod;
 use Ray\Di\Di\Assisted;
