+++ tests/NamedParameterTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace BEAR\Resource;
 
+use BEAR\Resource\Exception\NotBackedEnumException;
 use BEAR\Resource\Exception\ParameterEnumTypeException;
 use BEAR\Resource\Exception\ParameterException;
 use BEAR\Resource\Exception\ParameterInvalidEnumException;
